requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
scikit-learn
pandas
numpy
PyMuPDF>=1.24.0
sentence-transformers
faiss-cpu
google-generativeai
